refactor types folder
