+++ apiWrapper.tsx
@@ -1,5 +1,5 @@
 import { SearchResults, SearchResultsDataWrapper } from "./types/SearchResults";
-import { PullRequest } from "./types/PREntry";
+import { PullRequest } from "./types/PullRequest";
 import { graphql } from "@octokit/graphql";
 
 const fetchUserDetails = async (accessToken: string) => {
