+++ types/Author.tsx
@@ -1,0 +1,5 @@
+export interface Author {
+    avatarUrl: string;
+    login: string;
+    url: string;
+  }
