+++ types/PullRequestEntry.tsx
@@ -1,0 +1,5 @@
+import { PullRequest } from "./PullRequest";
+
+export interface PullRequestEntry {
+  node: PullRequest;
+}
