+++ types/Repository.tsx
@@ -1,0 +1,4 @@
+export interface Repository {
+    nameWithOwner: string;
+    description?: string;
+  }
